Separate tinymce and contentful coupling
