+++ src/tinymce/defaults.js
@@ -1,0 +1,10 @@
+const defaultMenubar = "edit view insert format tools hubspot";
+
+const defaultToolbar =
+  "hubspot codesample | undo redo | formatselect | bold italic | alignleft aligncenter alignright alignjustify | numlist bullist outdent indent";
+
+const defaultPlugins = [
+  "autoresize textpattern lists advlist media image code hubspot",
+];
+
+export { defaultMenubar, defaultToolbar, defaultPlugins };
