+++ src/tools/tweak.js
@@ -1,0 +1,10 @@
+export function tweak(param) {
+  var t = param.trim();
+  if (t === "false") {
+    return false;
+  } else if (t === "") {
+    return undefined;
+  } else {
+    return t;
+  }
+}
